Create base joint class
