+++ geoapps/inversion/joint/joint_cross_gradient/constants.py
@@ -21,6 +21,8 @@ inversion_defaults = {
     "topography": None,
     "group_a": None,
     "group_b": None,
+    "group_c": None,
+    "alpha_j": 1.0,
     "mesh": None,
     "output_tile_files": False,
     "inversion_style": "voxel",
@@ -33,15 +35,15 @@ inversion_defaults = {
     "max_line_search_iterations": 20,
     "max_cg_iterations": 30,
     "tol_cg": 1e-4,
-    "alpha_s": 1.0,
-    "length_scale_x": 1.0,
-    "length_scale_y": 1.0,
-    "length_scale_z": 1.0,
-    "s_norm": 0.0,
-    "x_norm": 2.0,
-    "y_norm": 2.0,
-    "z_norm": 2.0,
-    "gradient_type": "total",
+    "alpha_s": None,
+    "length_scale_x": None,
+    "length_scale_y": None,
+    "length_scale_z": None,
+    "s_norm": None,
+    "x_norm": None,
+    "y_norm": None,
+    "z_norm": None,
+    "gradient_type": None,
     "max_irls_iterations": 25,
     "starting_chi_factor": None,
     "f_min_change": 1e-4,
@@ -105,26 +107,121 @@ default_ui_json = {
     },
     "group_a": {
         "main": True,
-        "group": "Data",
+        "group": "Joint",
         "label": "Group A",
         "groupType": "{55ed3daf-c192-4d4b-a439-60fa987fe2b8}",
         "value": "",
     },
     "group_b": {
         "main": True,
-        "group": "Data",
+        "group": "Joint",
         "label": "Group B",
         "groupType": "{55ed3daf-c192-4d4b-a439-60fa987fe2b8}",
         "value": "",
     },
     "group_c": {
         "main": True,
-        "group": "Data",
+        "group": "Joint",
         "label": "Group C",
         "groupType": "{55ed3daf-c192-4d4b-a439-60fa987fe2b8}",
         "optional": True,
         "enabled": False,
         "value": "",
+    },
+    "alpha_j": {
+        "min": 0.0,
+        "group": "Joint",
+        "label": "Smallness weight",
+        "value": 1.0,
+        "main": True,
+        "lineEdit": False,
+        "tooltip": "Weight applied to the cross gradient regularization (1: equal weight with the standard Smallness and Smoothness terms.)",
+    },
+    "alpha_s": {
+        "min": 0.0,
+        "enabled": False,
+        "groupOptional": True,
+        "group": "Regularization",
+        "label": "Smallness weight",
+        "value": 1.0,
+        "tooltip": "*Overwrite*: Constant ratio compared to other weights. Larger values result in models that remain close to the reference model",
+    },
+    "length_scale_x": {
+        "min": 0.0,
+        "group": "Regularization",
+        "label": "X-smoothness weight",
+        "tooltip": "*Overwrite*: Larger values relative to other smoothness weights will result in x biased smoothness",
+        "value": 1.0,
+        "enabled": False,
+    },
+    "length_scale_y": {
+        "min": 0.0,
+        "group": "Regularization",
+        "label": "Y-smoothness weight",
+        "tooltip": "*Overwrite*: Larger values relative to other smoothness weights will result in y biased smoothness",
+        "value": 1.0,
+        "enabled": False,
+    },
+    "length_scale_z": {
+        "min": 0.0,
+        "group": "Regularization",
+        "label": "Z-smoothness weight",
+        "tooltip": "*Overwrite*: Larger values relative to other smoothness weights will result in z biased smoothess",
+        "value": 1.0,
+        "enabled": False,
+    },
+    "s_norm": {
+        "min": 0.0,
+        "max": 2.0,
+        "group": "Regularization",
+        "label": "Smallness norm",
+        "value": 0.0,
+        "precision": 2,
+        "tooltip": "*Overwrite*: Lp-norm [0, 2] measure applied to model values.",
+        "lineEdit": False,
+        "enabled": False,
+    },
+    "x_norm": {
+        "min": 0.0,
+        "max": 2.0,
+        "group": "Regularization",
+        "label": "X-smoothness norm",
+        "value": 2.0,
+        "precision": 2,
+        "tooltip": "*Overwrite*: Lp-norm [0, 2] measure applied to model EW-gradient values.",
+        "lineEdit": False,
+        "enabled": False,
+    },
+    "y_norm": {
+        "min": 0.0,
+        "max": 2.0,
+        "group": "Regularization",
+        "label": "Y-smoothness norm",
+        "value": 2.0,
+        "precision": 2,
+        "tooltip": "*Overwrite*: Lp-norm [0, 2] measure applied to model NS-gradient values.",
+        "lineEdit": False,
+        "enabled": False,
+    },
+    "z_norm": {
+        "min": 0.0,
+        "max": 2.0,
+        "group": "Regularization",
+        "label": "Z-smoothness norm",
+        "value": 2.0,
+        "precision": 2,
+        "tooltip": "*Overwrite*: Lp-norm [0, 2] measure applied to model vertical-gradient values.",
+        "lineEdit": False,
+        "enabled": False,
+    },
+    "gradient_type": {
+        "choiceList": ["total", "components"],
+        "group": "Regularization",
+        "label": "Gradient type",
+        "value": "total",
+        "tooltip": "*Overwrite*: Type of model gradients used by the Lp-norm measures.",
+        "verbose": 3,
+        "enabled": False,
     },
     "tile_spatial": None,
 }
