+++ geoapps/inversion/joint/joint_surveys/params.py
@@ -9,9 +9,7 @@ from __future__ import annotations
 
 from copy import deepcopy
 
-from geoh5py.groups.simpeg_group import SimPEGGroup
-
-from geoapps.inversion.params import InversionBaseParams
+from geoapps.inversion.joint.params import BaseJointParams
 
 from .constants import (
     default_ui_json,
@@ -23,12 +21,12 @@ from .constants import (
 )
 
 
-class JointSurveysParams(InversionBaseParams):
+class JointSurveysParams(BaseJointParams):
     """
     Parameter class for gravity->density inversion.
     """
 
-    PHYSICAL_PROPERTY = ""
+    _physical_property = ""
 
     def __init__(self, input_file=None, forward_only=False, **kwargs):
         self._default_ui_json = deepcopy(default_ui_json)
@@ -36,38 +34,22 @@ class JointSurveysParams(InversionBaseParams):
         self._forward_ui_json = deepcopy(forward_ui_json)
         self._inversion_defaults = deepcopy(inversion_defaults)
         self._inversion_ui_json = deepcopy(inversion_ui_json)
-        self._inversion_type = "joint single property"
+        self._inversion_type = "joint surveys"
         self._validations = validations
-        self._group_a = None
-        self._group_b = None
-        self._group_c = None
-        self._out_group = None
 
         super().__init__(input_file=input_file, forward_only=forward_only, **kwargs)
 
     @property
-    def group_a(self):
-        """First SimPEGGroup inversion."""
-        return self._group_a
+    def physical_property(self):
+        """Physical property to invert."""
+        return self._physical_property
 
-    @group_a.setter
-    def group_a(self, val: SimPEGGroup):
-        self.setter_validator("group_a", val, fun=self._uuid_promoter)
-
-    @property
-    def group_b(self):
-        """Second SimPEGGroup inversion."""
-        return self._group_b
-
-    @group_b.setter
-    def group_b(self, val: SimPEGGroup):
-        self.setter_validator("group_b", val, fun=self._uuid_promoter)
-
-    @property
-    def group_c(self):
-        """Third SimPEGGroup inversion."""
-        return self._group_c
-
-    @group_c.setter
-    def group_c(self, val: SimPEGGroup):
-        self.setter_validator("group_c", val, fun=self._uuid_promoter)
+    @physical_property.setter
+    def physical_property(self, val: list[str]):
+        unique_properties = list(set(val))
+        if len(unique_properties) > 1:
+            raise ValueError(
+                "All physical properties must be the same. "
+                f"Provided SimPEG groups for {unique_properties}."
+            )
+        self._physical_property = unique_properties[0]
